Move Solido logic into separate module
