+++ projects/lido/index.js
@@ -1,9 +1,8 @@
 const axios = require('axios');
 const abis = require('./abis.json')
 const sdk = require('@defillama/sdk')
-const { Connection, PublicKey } = require('@solana/web3.js');
-const { deserializeUnchecked } = require('borsh');
-const {Lido, schema} = require('./Lido')
+const { Connection } = require('@solana/web3.js');
+const sol = require('./sol-helpers')
 
 const ethContract = '0xae7ab96520de3a18e5e111b5eaab095312d7fe84';
 
@@ -33,32 +32,13 @@ async function eth(timestamp, ethBlock, chainBlocks) {
 
 async function solana(timestamp, ethBlock, chainBlocks) {
   const connection = new Connection('https://solana-api.projectserum.com/');
-  const validatorsBalance = await retrieveValidatorsBalance(connection)
-  const reserveAccountBalance = await retrieveReserveAccountBalance(connection)
+  const validatorsBalance = await sol.retrieveValidatorsBalance(connection)
+  const reserveAccountBalance = await sol.retrieveReserveAccountBalance(connection)
 
   const totalSolInLamports = validatorsBalance + reserveAccountBalance;
   return {
     'solana': totalSolInLamports/1e9
   }
-}
-
-async function retrieveValidatorsBalance(connection) {
-  const accountInfo = await connection.getAccountInfo(new PublicKey("49Yi1TKkNyYjPAFdR9LBvoHcUjuPX4Df5T5yv39w2XTn"));
-  const deserializedAccountInfo = deserializeUnchecked(
-    schema,
-    Lido,
-    accountInfo.data,
-  );
-  return deserializedAccountInfo.validators.entries
-    .map(pubKeyAndEntry => pubKeyAndEntry.entry)
-    .map(validator => validator.stake_accounts_balance.toNumber())
-    .reduce((prev, current) => prev + current, 0)
-}
-
-async function retrieveReserveAccountBalance(connection) {
-  const accountInfo = await connection.getAccountInfo(new PublicKey("3Kwv3pEAuoe4WevPB4rgMBTZndGDb53XT7qwQKnvHPfX"));
-  const rent = await connection.getMinimumBalanceForRentExemption(accountInfo.data.byteLength);
-  return accountInfo.lamports - rent;
 }
 
 module.exports = {
